requirements.txt
pip==25.3
python-decouple==3.8
django==5.2.8
pytest==9.0.1
requests==2.28.2
psycopg2-binary